chore(origin-backend): reorganize the folders and files in a more readable manner
